// app/types/getstarted.tsx
export interface GetStartedSettings {
    id: number;
    background_color_start: string;
    background_color_mid: string;
    background_color_end: string;
    title_prefix: string;
    title_italic: string;
    title_suffix: string;
    button_text: string;
    button_background_color: string;
    button_text_color: string;
    button_hover_color_start: string;
    button_hover_color_end: string;
    calendar_event_title: string;
    calendar_event_details: string;
    calendar_event_location: string;
    calendar_event_email: string;
    created_at: string;
    updated_at: string;
  }